fix(cleaning-schedule): move controllers to controller sub-package, apply project conventions
