+++ development/service-ops-api/src/main/java/com/thundercrew/opsapi/cleaningschedule/controller/CleaningScheduleReadController.java
@@ -1,12 +1,12 @@
-package com.thundercrew.opsapi.cleaningschedule;
+package com.thundercrew.opsapi.cleaningschedule.controller;
 
 import com.thundercrew.opsapi.cleaningschedule.dto.CleaningScheduleReadResponse;
 import com.thundercrew.opsapi.cleaningschedule.service.CleaningScheduleQueryService;
+import java.util.List;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import java.util.List;
 
 @RestController
 @RequestMapping("/api/v1/cleaning-schedules")
@@ -19,8 +19,8 @@ public class CleaningScheduleReadController {
     }
 
     @GetMapping
-    public List<CleaningScheduleReadResponse> list(
-        @RequestParam(required = false) String bikeId
+    List<CleaningScheduleReadResponse> list(
+            @RequestParam(required = false) String bikeId
     ) {
         if (bikeId != null) return queryService.findByBikeId(bikeId);
         return queryService.findAll();
